Fix circular import for permission_checks and move is_staff to .checks
is_staff was imported from resources.utils before, but it had been in the resources/checks/ directory, so it makes more sense to have it be imported from resources.checks.__init__ as well.
