+++ resources/checks/__init__.py
@@ -6,10 +6,12 @@ __all__ = [
     'InsufficientPermissionsCheckFailure',
     'CommandDoesNotSupportDMsCheckFailure',
     'ModuleNotEnabledCheckFailure',
-    'MissingAttributesCheckFailure'
+    'MissingAttributesCheckFailure',
+    'is_staff',
+    'is_admin',
 ]
 
-
+from .permissions import is_staff, is_admin
 from .permission_checks import is_admin_check, is_staff_check
 from .command_checks import not_in_dms_check, module_enabled_check
 from .errors import (
